requirements.txt
pandas==1.1.2
requests-html==0.10.0
PyPDF2==1.26.0
tabula-py